requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
langgraph==0.2.45
langchain==0.3.7
langchain-openai==0.2.8
google-auth==2.23.0
google-auth-oauthlib==1.1.0
google-api-python-client==2.108.0
python-dotenv==1.0.0
pydantic==2.9.0
tenacity==8.5.0
langsmith==0.1.147
python-json-logger==2.0.7
cachetools==5.3.2
email-validator==2.1.0